feat(rust): ADR-168 S0 parity — Rust CI + npm health bridge (#11)
Advance cli/rust-s0-scaffold toward ts_deleted with production-grade
parity proof matching the groundatlas S0 playbook:

- Add Rust CI job (cargo test + clippy -D warnings)
- Wire MEDIA_CURATOR_RUST_CLI=1 npm delegation bridge (src/external/rustCli.ts)
- Add --health fast-path bootstrap to avoid loading pipeline deps
- Add bridge integration tests (tests/rustCli.bridge.test.ts)

TS pipeline authority retained; S1 discovery/gather fixture parity next.
